Moved create_algo to primary_algo.py
